Add/fix tests for new shadow stack allocation scheme.
