+++ tests/c/guard_consting.c
@@ -4,7 +4,7 @@
 //   stderr:
 //     ...
 //     %{{30}}: i1 = sgt %{{_}}, 0i32
-//     guard true, %{{30}}, [{{0}}:%{{0_2}}: %{{0}}, {{0}}:%{{0_6}}: %{{1}}, {{0}}:%{{0_7}}: %{{2}}, {{0}}:%{{0_8}}: %{{3}}, {{0}}:%{{0_9}}: %{{4}}, {{0}}:%{{9_1}}: 0i1] ; ...
+//     guard true, %{{22}}, [{{0}}:%{{0_8}}: %{{0}}, {{0}}:%{{0_9}}: %{{1}}, {{0}}:%{{0_10}}: %{{2}}, {{0}}:%{{0_11}}: %{{3}}, {{0}}:%{{9_1}}: 0i1] ; ...
 //     ...
 
 // Check that if a guard's life variables include the condition operand, that
