+++ tests/c/simple_peeling.c
@@ -8,16 +8,16 @@
 //     ...
 //     header_start ...
 //     ...
-//     header_end [%{{0}}, %{{1}}, %{{2}}, %{{3}}, %{{4}}, %{{25}}]
+//     header_end [%{{0}}, %{{1}}, %{{2}}, %{{3}}, %{{4}}]
 //     ...
-//     body_start [%{{30}}, %{{31}}, %{{32}}, %{{33}}, %{{34}}, %{{35}}]
+//     body_start [%{{19}}, %{{20}}, %{{21}}, %{{22}}, %{{23}}]
 //     ...
 //     body_end ...
 //     ...
 //     --- End jit-post-opt ---
 //     ...
 
-// Check that basic trace compilation works.
+// Check that basic trace peeling works.
 
 #include <assert.h>
 #include <stdio.h>
